feat: Added custom colors and examples
