+++ src/routes/theme/+page.svelte
@@ -1,0 +1,29 @@
+<div>
+    <h1>H1 Element (h3 in figma)</h1>
+    <h2>H2 Element (h4 in figma)</h2>
+    <h3>H3 Element (list in figma)</h3>
+    <div class="text-placeholder">
+        Placeholder Element (subtle-medium in figma + italic + slate/500)
+    </div>
+    <div class="text-hint">Hint Element (detail in figma + neutral/500)</div>
+    <div class="text-default">Default Element (subtle in figma)</div>
+    <div class="text-default-m">Default Element (subtle-medium in figma)</div>
+    <div class="text-default-sb">Default Element (subtle-semibold in figma)</div>
+</div>
+<div class="bg-gray-800 py-1">
+    <div class="text-decline-red">This is a decline-red colored text</div>
+    <div class="text-maybe-yellow">This is a maybe-yellow colored text</div>
+    <div class="text-accept-green">This is an accept-green colored text</div>
+    <div class="text-warning-amber">This is a warning-amber colored text</div>
+</div>
+
+<div class="p-3 bg-white w-fit">
+    <div class="border-container-border-grey border rounded-md p-2 shadow-lg">
+        <div
+            class="border-input-border-slate border w-fit p-2 m-2 rounded-md text-input-placeholder-slate"
+        >
+            This is a container with an input-border-slate border and an input-placeholder-slate
+            text inside a container with a container-border-grey border (and shadow)
+        </div>
+    </div>
+</div>
